requirements.txt
fastapi
uvicorn[standard]
tensorflow-cpu
pillow
numpy
python-multipart
pydantic